+++ files/gitlab-cookbooks/gitlab/recipes/mattermost.rb
@@ -114,52 +114,6 @@ template config_file_path do
   notifies :restart, "service[mattermost]"
 end
 
-##################
-# Upgrade from V2 to V3 workarounds
-backup_done = node['gitlab']['mattermost']['db2_backup_created']
-default_team_name_for_v2_upgrade = node['gitlab']['mattermost']['db2_team_name']
-default_team_name_set = !default_team_name_for_v2_upgrade.nil?
-log_file = File.join(mattermost_log_dir, "mattermost.log")
-mattermost_helper = MattermostHelper.new(node, mattermost_user, mattermost_home)
-
-# If mattermost version returns exit status different than 0, database
-# migration most likely is not possible
-# stop the running service, something went wrong
-execute "/opt/gitlab/bin/gitlab-ctl stop mattermost" do
-  retries 20
-  only_if { mattermost_helper.version.nil? }
-end
-
-if backup_done && default_team_name_set
-  execute "/opt/gitlab/bin/gitlab-ctl start mattermost" do
-    retries 2
-    only_if { mattermost_helper.version.nil? && MattermostHelper.upgrade_db_30(config_file_path, mattermost_user, default_team_name_for_v2_upgrade) == 0 }
-  end
-end
-
-bash "Show the message of the failed upgrade." do
-  code <<-EOS
-    echo "!!!!Automatic database upgrade failed.!!!\n
-    If you are upgrading from Mattermost v2 to v3
-    make sure that you have backed up your database
-    and then in /etc/gitlab/gitlab.rb set:
-
-    mattermost['db2_backup_created'] = true
-    mattermost['db2_team_name'] = \"TEAMNAME\"\n
-
-    where "TEAMNAME" is the name of the default team.
-    Run gitlab-ctl reconfigure again.
-    See: \n
-    https://docs.gitlab.com/omnibus/gitlab-mattermost/#upgrading-gitlab-mattermost-from-versions-prior-to-8-9 \n
-    for more information.\n
-    " >> #{log_file}
-  EOS
-  user mattermost_user
-  only_if { mattermost_helper.version.nil? && !(backup_done && default_team_name_set) }
-end
-
-###############
-
 ###
 # Mattermost control service
 ###