+++ CHANGELOG.md
@@ -10,12 +10,13 @@ omnibus-gitlab repository.
 - Bump gitlab-elasticsearch-indexer version to 0.2.0 bba8edd3
 - Adjust various default values for PostgreSQL based on GitLab.com
 
-
 9.3.3
 
 - Allow sidekiq-cluster to run without having sidekiq enabled
+- Remove outdated Mattermost v2 DB upgrade code
 
 9.3.2
+
 - Update gitlab-monitor to 1.8.0
 
 9.3.1